hwid: Allow to specialize the status of the region components.
The region component in a HWID database is auto generated according to
the system regions.  Therefore, the HWID unittests might become fail if
a region is removed from the confirmed list.  And we didn't have a way
to fix the database because of the backport compatibility.

This CL allows users to customize the region components by specifying
the status of some regions.  For example:

  regions: !region_component
    unsupported: [aa]
    deprecated: [bb, cc]

This marks the region code `aa` to `unsupported` status and `bb`, `cc`
to `deprecated` status.

BUG=b:110596222
TEST=make test

Change-Id: If75e5ad393275f36f74ff688b258b2b75d2a6c27
Reviewed-on: https://chromium-review.googlesource.com/1160080
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Marco Chen <marcochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2852309
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>