pytests/shopfloor_service: Upload the test results.
The original shopfloor_service pytest was designed to let the caller
specify the arguments.  However, it's impossible (or super complex)
to set the test path and the test status as the argument for
shopfloor_service in JSON based test list.  So this CL allows the
pytest can obtain the test status to upload by itself.

BUG=b:79621937
TEST=manually test

Change-Id: Ifcfeab54e81e3f47eb25f211b158324fc086d1e8
Reviewed-on: https://chromium-review.googlesource.com/1198926
Reviewed-by: Yong Hong <yhong@google.com>
Commit-Queue: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1248329