setup: Add RMA_AUTORUN in RMA shim.
Add a flag RMA_AUTORUN in lsb-factory in RMA shim. If autorun is
enabled, RMA shim will do RSU if HWWP is on, and will do install if
HWWP is off. When creating the RMA shim, autorun is default enabled.

Same as CL:1322469 but for factory branches without image_tool.

BUG=b:118785701
TEST=run `edit_lsb_factory` on an RMA shim and change RMA_AUTORUN
     option.

Change-Id: I4f8b5fe076dd3c6a26944a6c49f8c1fde823ae23
Reviewed-on: https://chromium-review.googlesource.com/c/1375182
Reviewed-by: Marco Chen <marcochen@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>