gooftool: verify_system_time: don't fail in rma mode
There is no factory server in RMA center, so the device cannot get
correct time.  And there might be a case that the device time is always
before release image file system creation time.

Test image and factory install shim are both came from factory branch,
which is not as active as release branch, so /etc/lsb-release and
/usr/local/etc/lsb-factory creation time will be earlier than release
image time.  If the device time was earlier than release image creation
time when installing RMA shim, then this earlier time will persist
during testing, and the test will fail.  (e.g. the test image, toolkit,
install shim were built at mid 2018, while release image was built at
early 2019, and the device local time was late 2018).

We had some experiment and found that the device can update to correct
time if it has internet access (I set the device to 1970/01/01 w/o
connecting to internet, reboot, the device is still at 1970.  And then
connect the device to the internet, a few minutes later, the device has
correct time).

Therefore, a workaround for the RMA case is, if the current time is
earlier than release image creation time, we set the current time to
release image creation time and continue.

This is enabled in RMA mode only, because in factory, we should have a
good enough source of time (factory server), and we want the date
related VPD values (rlz embargo end date) to be somewhat accurate.

BUG=None
TEST=manual on device

Change-Id: Icfb96dc447cb267beb1ce887ce7784c348d53615
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1421538
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>