hwid: Remove pprint when printing probe result.
pprint adds newlines when printing, causing the error message in goofy
console unclear because it only captures the last line of the output. It
causes more debug time in repair centers. This CL removes pprint and
prints the error message in a single line, so that the logging system
can capture the whole message.

BUG=None
TEST=manually test on DUT

Change-Id: I567548576273961ef0fc42679d2884bb130b20de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1488595
Reviewed-by: Marco Chen <marcochen@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>