hwid: add new status 'duplicate'
In CL:1195191, we have covered the case that a component is subset of
other component, if regular expression is not used.  This should have
covered most cases.

To cover the regular expression case, for example:

  comp_1:
    values:
      part_number: PART_NUMBER_1-1
  comp_2:
    values:
      part_number: PART_NUMBER_1-2

And we later found that, PART_NUMBER_1-1 and PART_NUMBER_1-2 are
actually the same component in AVL database, 'PART_NUMBER_1', with some
suffix ('-1' and '-2') that we actually don't care.  The following
change [1] should not work:

  comp_1:
    values:
      part_number: !re PART_NUMBER_1-.*
  comp_2:
    values:
      part_number: PART_NUMBER_1-2

Because comp_2 is a subset of comp_1.  (There is a bug in current code,
so the encoder won't complain, but it is wrong).  To solve this kind of
ambiguity, status 'duplicate' is introducted.

  comp_1:
    values:
      part_number: !re PART_NUMBER_1-.*
  comp_2:
    status: duplicate
    values:
      part_number: PART_NUMBER_1-2

In this case, encoder will ignore comp_2.  And comp_2 can still be used
for decoding.  Also fix the bug in cros.factory.hwid.v3.probe, so the
change like [1] will fail during HWID encode.

BUG=b:204009593
TEST=py/hwid/v3/base32_unittest.py
TEST=py/hwid/v3/identity_unittest.py
TEST=py/hwid/v3/base8192_unittest.py
TEST=py/hwid/v3/probe_unittest.py
TEST=py/hwid/v3/builder_unittest.py
TEST=py/hwid/v3/rule_unittest.py
TEST=py/hwid/v3/converter_unittest.py
TEST=py/hwid/v3/transformer_unittest.py
TEST=py/hwid/v3/database_unittest.py
TEST=py/hwid/v3/hwid_cmdline_unittest.py
TEST=py/hwid/v3/verifier_unittest.py
TEST=py/hwid/v3/hwid_rule_functions_unittest.py
TEST=py/hwid/v3/yaml_wrapper_unittest.py
TEST=py/hwid/v3/hwid_utils_unittest.py

Change-Id: Ibc8938d0308930d5fbb23d138fdf117e46f0b6b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2852313
Reviewed-by: Yong Hong <yhong@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3245012
Commit-Queue: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>