pytests/camera: Uses `<video>` element to capture frames.
In e2e mode, the pytest used to capture frames by the JavaScript
API `ImageCapture.getFrame()`.  However, starts from
test image 11893.0.0, that API becomes not working if the
media stream instance to capture is not set to `srcObject` of a
`<video>` element.  Therefore we has to create an additional
`<video>` element.  And because we can directly capture an frame
from that element, we no long need to invoke `ImageCapture.getFrame()`
API.

BUG=b:129033034
TEST=manual test

Change-Id: I832b01a5fc30639c19ace71e1a6ee6576d340fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1549224
Reviewed-by: Yong Hong <yhong@google.com>
Commit-Queue: Yong Hong <yhong@google.com>
Tested-by: Corner Li <cornerli@google.com>