hwid: configless_field: can ignore dram and storage in rma mode
RMA spare boards in factory can pre-generate HWID string, and encode
with special encoding pattern 'RMA'.  In this case, components not in
the list of encoded fields will be omitted.  And DRAM (memory), storage
might be one of omitted fields.  In this case, it is okay to omit them
in configless fields too.

BUG=b:133126697
TEST=make test; manual test

Change-Id: I87c2a76da6c35d3f393698645a5f940998506c97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1617148
Reviewed-by: Wei-Han Chen <stimim@chromium.org>
Commit-Queue: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>