camera: Restart cros-camera when failed to grabFrame().
This is a temporary workaround of an unknown race in Chromium that is
consistenly happened on Atlas when running camera stress test.

The real culprit is hard to debug because we do not have a reliable way
to reproduce the race within a short period of time.  Add this hack to
unblock the factory process because it's a software issue.

BUG=b:133810965
TEST=Inject some random errors in grabFrame() manually, and check the
     stress test is still running after camera service is restarted.

Change-Id: I5ba8bb36f06f126b1caba939e9b036925be45035
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1647103
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>