update_cr50_firmware: always block prePVT Cr50 FW in PVT stage
PrePVT Cr50 firmware has a higher version than the prod one, which means
it's impossible to 'upgrade' to prod Cr50 image once a prePVT image is
flashed, and this could be a disaster if happended in mass production.

To prevent any potential misconfiguration in test lists bypass the check
in PVT/MP phase, adding a checker to always fail the task if (prePVT
Cr50 image, PVT phase) combination is found.

BUG=b:153277262
TEST=manual on Krane;
     Task fails if prePVT image is provided in PVT phase or phase is
     prior to PVT but skip_prepvt_flag_check is unset.

Change-Id: I86f640e006c6e651deef1b621b94f98cc257184d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2141524
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>