hwid/v3: align ListFiles behavior
Previously CloudStorageAdapter.ListFiles could filter filename prefix
with the prefix argument and return a list of `GCSFileStat` instances.
This implementation will be difficult to abstract with
LocalFilesystemAdapter.  In this CL we change the behavior to align
LocalFilesystemAdapter which uses os.listdir to list filenames under the
folder.

BUG=b:155367164
TEST=${FACTORY_REPO}/deploy/cros_hwid_service.sh test

Change-Id: I929584812fdb7e4824dce9217cceb0c78eabf9e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2184011
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>