hwid/service: allow partial vp generator for a category
The meaning of each component category in HWID DB might not be
identical to the meaning in runtime probe.  For example, the category
`video` includes both cameras and video codecs in HWID.  But runtime
probe separates the type `camera` out.

This CL revises the HWID service to support such flexibility:

- Since runtime probe doesn't support video codec so far, having a
  `video` component in HWID DB without any suitable probe statement
  generator becomes acceptable.

- Whether a component should be included by the dutlabels now is
  determined by whether the verification payload (vp) generator
  takes that component as an input.

BUG=b:160120501
TEST=make test; ./deploy/cros_hwid_service.sh test

Change-Id: I9a311d956d6a678049f09615f6c9badd69ae82dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2276076
Commit-Queue: Yong Hong <yhong@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>