audio_loop: Allow longer duration in sinewav test
Previously, we didn't check if the duration is longer than the wav file.
This CL adjusts the length of the wav file to be always longer than the
duration and fix BaseAudioControl._GetPIDByName if no process presents.

Also shows a more clear log when aplay ends before we call
_StopPlaybackWavFile.

BUG=b:150912730
TEST=factory run SMT.AudioJack

Change-Id: I8e968dbb9a24651ca27c4a42549830b1316cba3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2879129
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>