audio: Detect 2mic ucm config
Previously we need to define device_map in device API so as to use 2mic
ucm config. But there are boards with 1mic and 2mic variants and we can
not make them use the same device map. This CL uses 'list _devices' in
alsaucm to determine the dut is 1mic or 2mic.

BUG=b:150912730
TEST=factory run SMT.SpeakerDmic on 2mic dut with ucm config

Change-Id: If55ca06c5deb3ccd25c9e2e7230ef968a764e1ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2879128
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>