camera: Draw face recognition rectangle on the frontend
For the face detection part, draw rectangle on the frontend instead of
the backend. This removes the need of transmitting the whole image back
to frontend, which slightly improves the fps in the face detection part
in e2e mode.

This also changes the HTML layout of the camera test, and makes the
camera view fills the whole area when show_image is set to true. Since
this is resized by Chrome using CSS, should have little impact on
performance.

BUG=None
TEST=manually test the camera tests.

Change-Id: I8a91eca173c6fd9fe5ba86e1fbf7a5c8bf1ecee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2409680
Tested-by: Fei Shao <fshao@chromium.org>
Auto-Submit: Fei Shao <fshao@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>