removable_storage: Fix test when skip_insert_remove is true
When the skip_insert_remove is true, the test is stuck in
GetDeviceNodeBySysPath because it cannot find the device. Using the
regex to match can solve this.
The GetUsbSpeed also fails because the target device is not a pyudev
Device object, so we will use the pyudev Device.

BUG=b:172533897
TEST=run removalbe_storage.py on DUT

Change-Id: I3bb336c0a02db23d7d44156d4c717eb76837e242
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2546246
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
(cherry picked from commit dec8baddaeefd0290ecbc691129cb5330e62216e)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2550086
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>