Instalog: output_factory_report: Download multiple archives in parallel
Because the performance of the output_factory_report plugin is not good
enough, we change the structure of the plugin and implement parallelly
download.
Before this CL, we need to wait till an archive is downloaded. This
implementation can download the next few archives before the plugin
finish an archive.

BUG=b:189805423
TEST=manually test

Cq-Depend: chromium:3826457
Change-Id: I684fbaa1e785e47b4b24807c3c83ebdd2483b26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3195282
Tested-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Chun-Tsen Kuo <chuntsen@chromium.org>