custom_type: get custom_type from model_sku.json
We add custom_type in model_sku.json, that is used to identify
whitelabel and rebrand projects. We read the information in
gooftool and check whether the whitelabel filed in VPD doesn't
exist only for whitelabel projects.

Please refer to b:175293581 for more details.

BUG=b:175293581
TEST=make test

Change-Id: I8d052ea7f023d3213eb981957b062fc3d9006137
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3089004
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Isaac Lee <isaaclee@google.com>
Commit-Queue: Isaac Lee <isaaclee@google.com>