unittest: fix resource warning in net_utils
The implementation of TimeoutXMLRPCTransport in net_utils does not store
the connection for close if exception occurs. Store the connection which
matches the parent class to close the connection when exception occurs.
The unit test for net_utils does not close the server and socket
appropriately, fix it with the right API. Also, remove tests affected by
net_utils from the skip check list.

BUG=b:205776738
TEST=make test

Change-Id: Icb9586d37355659dfccf9c614675253b9e5b9133
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3287151
Reviewed-by: Chung-Sheng Wu <chungsheng@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>