unittest: fix resource warning in output http
The plugin output http opens file without for uploading but does not
close them after use. Close them after read the content and pass to
requests for uploading. The unit test of it does not close the server
after use, close it with the correct API and remove it from the resource
warning skip check list.

BUG=b:205776738
TEST=make test

Change-Id: Ieb216d65c886d0fb70d37d7871f9de8dc1bdf049
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3304465
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Chung-Sheng Wu <chungsheng@chromium.org>