smart_amp_utils: add regex to parse sound-card-init-conf
We add regex to parse the content of the sound-card-init-conf. We
also add new gooftool command `gooftool get_smart_amp_info` to get
the information about the smart amp. User is able to get the
channels of the smart amp from this new command.

BUG=b:209373189
TEST=run `gooftool get_smart_amp_info` on brya

Change-Id: I30afa5fc56026eee2a8c47cb3fec4ad5a1c67195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3319041
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>