alsa: Stop showing exception when ucm config exists
Previously, when the json audio configuration did not exist, we would
display an exception. However, when we have ucm configuration, the lack
is normal. The exception in the log makes the users think that the test
tool is wrong.

Only show exceptions when neither configuration exists.

BUG=b:209934813
TEST=(1) Run SpeakerDMic test when an invalid audio_conf is set.
     (2) See "Json audio config {config_name} is not valid. Fallback to"
         "UCM audio config." in the factory.log.

Change-Id: Ibbd53afd30f7cf01c6cb953e534ba4075ed0c605
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3331871
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>