presubmit_format: don't pass excluded files to yapf
Previously, we use `--exclude` argument of yapf to exclude files we
don't care.  However, when we change a file that should be excluded, we
would still pass the file to yapf.  And, because we create individual
yapf processes for each changed file, the yapf process that is passed
with an excluded file will exit with non-zero return code, because the
only input file it got is excluded.

  > yapf --exclude <pattern> <an excluded file>
  < yapf: Input filenames did not match any python files

In this CL, we changed the logic, we use fnmatch to filter out the files
we want to exclude, so they are never passed to yapf.

BUG=None
TEST=make test
TEST=make format

Change-Id: If042b2291d218ec2070eea30644c8627bcf2fee5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3347315
Tested-by: Stimim Chen <stimim@chromium.org>
Auto-Submit: Stimim Chen <stimim@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>