probe_info_service/protorpc_utils: packs error detail pb message in resp
It revises `protorpc_utils.py` so that the flask HTTP handler packs
the given protobuf message of the error detail into the specific
header in the HTTP response.  The RPC framework will further propagate
the data so that the client can seamlessly get the detail of the error.

BUG=None
TEST=manual test

Change-Id: I2565c4c13c41207b04559feec1032c630dff959c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3347317
Reviewed-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Liang-Chieh Chen <jamesqaq@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>