hwid/service: a stopgap to display DB checksum in bundle script
Previous version of HWID Bundle requires a line showing the checksum of
the DB in order to check if DUT needs to update this DB.  As we move the
bundle building flow to bundle_builder which compresses the files and
causes the HWID check to fail.  This CL introduces a stopgap to embed a
line showing the required line of checksum to pass the check.

BUG=b:211957606
TEST=deploy/cros_hwid_service.sh test

Change-Id: I36927586b27a968d99ba05a697d5a20ef3e5550d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3356520
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>