test_list: redefine MRCCache in run in stage
We use `warmup_post_shutdown` to avoid race condition in reboot during
run in stage, but `RunIn.MRCCache.RebootStep` is not defined in
`generic_run_in` so it is not be affected.  Define a new test group
`RunIn.RunInMRCCache` to replace `RunIn.MRCCache` in `generic_run_in`
to enable this option.

BUG=b:203168950, b:201251516
TEST=make test

Change-Id: I63fe46cd8a9242e40ad3974065d7b4438dcddbcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3356542
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Yu-An Wang <wyuang@google.com>