pytest: Use MLB serial in ProvisionDRMKey if possible
The `serial_number` VPD is not available when we build replacement
MLBs for RMA, so it's better to use `mlb_serial_number` if possible.
Use a random string to request the keybox if neither is present, as
having a serial number is not a strong requirement to request a keybox
in the factory.
This CL also moves the patches of `OEMCryptoClient` to the `setUp`
function to make the code cleaner.

BUG=b:213031131
TEST=make test
TEST=Run ProvisionDRMKey pytest

Change-Id: Ife6745134f51a631ef0e947bfd1c5fc317131b3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3365150
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Chih-Yao Chuang <jasonchuang@google.com>