generic_common: add cr50_smt_write_flash_info
The test item Cr50WriteWhitelabelFlags is renamed to
Cr50SMTWriteFlashInfo.  The pytest is updated as well.

The test item now accepts 3 arguments:
1. mlb_mode: is this MLB going to leave current factory? Such as LOEM
   project or RMA spare parts.
2. rma_mode: is the current factory proecess in RMA mode? This implies
   MLB mode for this test item.
3. enable_zero_touch: The test will try to enable ZTE if this is set.
   However, ZTE is always disabled if rma_mode is set.

If the current project is NOT whitelabel, and neither mlb_mode nor
rma_mode are set, the test does nothing. This is the current behavior.

Otherwise, the arguments will be passed to `gooftool
cr50_smt_write_flash_info`. The command will set cr50 flash info, to
make sure board ID is set with:
1. Partial value if this is a whitelabel project.
2. Full value if this is not a whitelabel project.

For backward compatibility, we keep Cr50WriteWhitelabelFlags as an alias
to Cr50SMTWriteFlashInfo.  The alias will be removed when we clean up
the references in the private overlay.

BUG=b:219670435
TEST=make test

Change-Id: I86eac03c311c63c32a72e8fa3bfb234bb51c877f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3449973
Tested-by: Stimim Chen <stimim@chromium.org>
Auto-Submit: Stimim Chen <stimim@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3691307
Tested-by: Jacob Chang <jacobrl_chang@compal.corp-partner.google.com>
Reviewed-by: Jacob Chang <jacobrl_chang@compal.corp-partner.google.com>