image_tool: Fix permission of the tempfile
When we open a tempfile, we cannot write it in another process from
another user even if we change the permission to 0o777.

Previously, we ask partner to replace `setup/image_tool ...` with
`sudo setup/image_tool ...` which is a workaround.

To prevent this, open a file in a temporary directory instead.

BUG=b:168680891
TEST=${SRC}/platform/factory/setup/image_tool rma create -f \
      --board brask \
      --factory_shim "${BUNDLE}/factory_shim/*.bin" \
      --test_image "${BUNDLE}/test_image/*.bin" \
      --release_image "${BUNDLE}/release_image/*.bin" \
      --hwid "${BUNDLE}/hwid/hwid*.sh" \
      --toolkit ${BUNDLE}/toolkit/install_factory_toolkit.run \
      --project_config "${BUNDLE}/project_config/*.tar.gz" \
      --no-firmware \
      --active_test_list main_brask \
      --output ${BUNDLE}/brask_rma_image.bin

Change-Id: Ib11518c7ad77cf8d82e7bade7e9a730e95865ea2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3467485
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
(cherry picked from commit 369f60baa248b7135753f8d2127c109d453a2367)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3479656
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>