hwid/v3: add entry points for setting internal flags
This CL adds a new argument `internal` for dump-related methods which
could dump HWID DBs with optional internal tags.

In addition, the checksum field in Database instances was only for
verification purpose but not used while dumping to file/text.
Therefore, a dummy string was used and the related methods are renamed
to Dump{File,Data}WithoutChecksum to prevent confusion in the future.

BUG=b:209362238
TEST=make test

Change-Id: I2542b3be456672e53bf5213aa26e4537fa8a3db4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3470219
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>