keyboard: Support more than 10 vivaldi keys
We allow more than 10 vivaldi keys now. Supports this in keyboard.py.
This CL is only used to unblock the factory test. The layout on the UI
still contains only 10 keys and extra keys are shown on the top left of
the keyboard. We plan to fix the UI in later changes.

BUG=b:221182587,b:222602376
TEST=Run the test with vivaldi_keyboard=true on a board with 13 vivaldi
     keys.

Change-Id: If03c9f87f2f0461e9e5eefe3639aa4fb0555b100
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3486942
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>