removable_storage: Set proper path for SD Performance test
If skip_insert_remove is True, we get the device path from
`GetDeviceNodeBySysPath` and it only accepts path as `/sys/block/sd*`.
But the path could be `/sys/block/mmcblk*` if connecting via an PCI
card reader. We need to accept both formats.

In addition, in order to trigger TimeoutError in sync_utils.WaitFor,
timeout for sync_utils.WaitFor is set to detect_timeout_secs, which
is less than the timeout for SD Performance test(timeout_secs).

BUG=b:219981042
TEST=make test
TEST=Run SD Performance test on DUT(volta)

Change-Id: I4cfe849d7622942360e00c22c13067cb7d4ced0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3500182
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Tzu-Min Sun <jimmysun@chromium.org>
Commit-Queue: Tzu-Min Sun <jimmysun@chromium.org>