edp_panel_timing: Add a pytest to ensure panel is with proper timing
The eDP panel driver 'panel-simple-dp-aux' is a driver on ARM for ePD
panel. It can drive unknown panels as well with generic but conservative
timings, which could impact user experiences such as boot time,
suspend/resume, etc. As panel-simple-dp-aux supports querying whether
the panel is driven with dedicated timing in debugfs, add a pytest in
the factory to ensure panel is driven with dedicated timings.

BUG=b:215432805
TEST=(1) test on DUT w/o eDP panel
     (2) test on DUT w/ eDP panel using hardcoded timing
     (3) test on DUT w/ eDP panel using detected timing
     (4) test on DUT w/ eDP panel using conservative timing

Change-Id: Ic182bb678c56358a8255344817a977ae48993587
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3514945
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>