py/hwid/service: respond whether entries are AVL sub-components
It appends a new boolean field called `is_subcomp` in `AvlInfo` message,
which indicates whether the component name indicates that the
corresponding AVL entry is a sub-component or not.

The flag value will always be `false` until the clients are upgraded
to handle the sub-component cases.

BUG=b:225963843
TEST=make test
TEST=./deploy/cros_hwid_service.sh test

Change-Id: Iae82a1ca0b09d7162cf4a2a584b00b5f40d645d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3552620
Reviewed-by: Clark Chung <ckclark@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>