hwid/service: add AVL-HWID probe value converter
This CL implements the converter classes which are used to convert the
AVL probe values from Probe Info Service to the key-value probe values
in HWID format.  In later CLs, the converted values will be used to
check the alignment status to enhance the decision making of CL
manipulation.

BUG=b:209362238
TEST=${FACTORY_REPO}/deploy/cros_hwid_service.sh test

Change-Id: I7a0e80ac5ee4bd3a569923c561eb9ca98d6ff7e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3641199
Commit-Queue: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>