hwid/service: add commit id information from HWID repo to bundle
As we add more logics (internal tags, reformat), the HWID DB in bundle
is not guaranteed to be identical to the one in chromeos-hwid repo.  We
should support the traceability of the source of the DBs extracted from
the bundle.  This CL adds the commit id in the bundle script as comment
which could link to the state of HWID DB in repo.

BUG=b:227427518
TEST=${FACTORY_REPO}/deploy/cros_hwid_service.sh test

Change-Id: I867c095bfc39db1284f70923f37a10907464bd6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3557883
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>