touchscreen: Add screen rotation functionality
The screen on DUT could have angle difference between the system and
event read from the evdev source, which causes inconsistency between
displayed UI instructions and detected input positions. This requires
the operator to first enter the tablet mode, then rotates the device for
compensating the angle difference to conduct this test. Add an
angle_compensation option to allow the pytest to automatically rotate
the UI before starting the pytest, and restores the rotation to auto
after the test.

BUG=b:228868896
TEST=Run touchscreen pytest on DUT w/o setting the angle_compensation

Change-Id: I4e2564b0607f4d611734042edd088f87bb850cec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3581532
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>