gooftool: whitelabel-tag is now custom-label-tag
On ToT, whitelabel-tag is renamed to custom-label-tag.  Even though the
VPD field `whitelabel_tag` is still allowed, the other parts in the code
base already renamed it to custom-label-tag.  This includes the
config.yaml in the release image.

In gooftool, we compares the identity in config.yaml of release image
and config.yaml of test image. If they don't match, the test will fail.
Therefore, we need to "normalize" the attribute name between release
image and test image, so they can match even though one is
`whitelabel-tag`, and the other one is `custom-label-tag`.

BUG=b:230051405
TEST=None

Change-Id: I9693f25652329427da2dba1cf61d782c74ff0950
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3601990
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>