gooftool: Fix doc string not pass within `Command` wrapper.
The Command decorator in gooftool does not update the output function
from the wrapped function, thus pydoc and argparse_utils cannot get the
original doc string. Use functools.wraps to update the wrapped Command
so that the doc string is updated to the output function, and users can
see the doc string when supplying the '-h' options for gooftool
sub-commands.

BUG=None
TEST=run `gooftool <sub-command> -h` for all sub-commands in gooftool
     and check command description shows up.

Change-Id: I4d1ffc2d590f6f3b0ceef34816d5ff38e7963aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3622654
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>