Instalog: Check number of processed reports matches to total reports
Check the number of processed reports matches to the total number of
reports in the archive. If the archive is a tar archive, check the
file names in it with `tar tvf`; if the archive is a zip archive, check
the file names in it with `unzip -l`. Log an error after finished
processing the archive if the number does not match to the number
reported from external tool.

BUG=b:232464200
TEST=make test

Change-Id: I04d169ed6d41603fbaa8925cf718b23424d45be9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3648859
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>