pytest: refactor keyboard.py
1. Rename down_keys to hold_keys
Down is the action that press the key, the action not release the
key is called hold.
2.Rename number_to_press to need_press_keys
We don't need to maintain the keys that are skipped.
This will be implemented in the next CL.
3. Merge all_keys into need_press_keys
need_press_keys is generated by all_keys, we don't need a set and
a dict with the same keys.
4. Handle order in OnKeyup
The intuition is raise if strict_sequential_press, otherwise do
nothing with the wrong order. It's a little complicated to implement
this logic in 2 functions by handling ignored_down_keys.
5. Maintain next_index
We can get next key needed to press in constant time instead of
loop in key_order_list.

BUG=b:232199509
TEST=test on device with arg allow_multi_keys, multi_keys_delay,
sequential_press, strict_sequential_press, repeat_times, has_numpad

Change-Id: Id1051adbdbb212116d5035f2a11db076046d14f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3645633
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>
Commit-Queue: Chih-Yao Chuang <jasonchuang@google.com>