py/probe: correctly probe devices from /sys/module/bluetooth/holders
A bug were introduced during code migration that, in
`generic_bluetooth.py`, the fallback approach to search devices in
`/sys/module/bluetooth/holders` might fail when it attempted to sort
the probe results, which are presented in dictionaries.

As sorting the probed results is not necessary, this CL simply makes
the probe function return the probed devices without caring the order.

BUG=b:234682160
TEST=manual test

Change-Id: Ia94172e31f23e85b1c4cd996a9f917df5483105f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3689098
Commit-Queue: Yong Hong <yhong@chromium.org>
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>