thunderbolt_loopback: Allow testing when multiple cards are inserted
The argument `controller_port` is used as the glob pattern of the target
controller path. Previously, the test is blocked whenever an non-target
controller path is found.

Now, we rename `controller_port` to `controller_patterns`, do not block
the test when a non-target controller is found, and only show the
non-target controllers in the log.

Also rename `devices` to `target_controllers` and `wrong_devices` to
`non_target_controllers` because it makes more sense that
`controller_patterns` are used to find controllers instead of devices.

BUG=b:234108363
TEST=run TBTLoopback01 and TBTLoopback11 on brya devices

Change-Id: Ib857be71230f4445bda4deefdbd141694e57cee9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3697686
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Phoebe Wang <phoebewang@chromium.org>