gooftool: auto propagate arguments of dependencies
Some higher level Gooftool commands depend on other Gooftool commands.
And this dependency can be multi-level.  For example,
- `finalize` depend on `cr50_finalize`
- `cr50_finalize` depend on `cr50_write_flash_info`

If we add an argument to `cr50_write_flash_info`, we usually want to
expose that argument via `finalize` and `cr50_finalize` too. In this
case, we need to add the new argument to all three functions.

To make it easier to manage the arguments, we make the arguments
available via: `Function.__args__`.

For example,
```
@Command('func_1',
         arg_x,
         arg_y)
def Function1(options):
  ...

@Command('func_2',
         *Function1.__args__,  # this will expand to arg_x, arg_y
)
def Function2(options):
  Function1(options)
```

In this case, when we add a new argument to Function1, we no longer need
to add it to Function2 manually.

BUG=None
TEST=make test
TEST=`gooftool finalize --help` shows the same set of arguments as
     before.

Change-Id: I47241af9bf08bee90532f5c8cfb80582f4153893
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/3901764
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>